+++ scripts/superdesk-legal-archive/controllers/index.js
@@ -1,0 +1,1 @@
+export { LegalArchiveController } from './LegalArchiveController';
