+++ scripts/superdesk-legal-archive/directives/index.js
@@ -1,0 +1,1 @@
+export { LegalItemSortbar } from './LegalItemSortbar';
