+++ scripts/superdesk-legal-archive/services/index.js
@@ -1,0 +1,1 @@
+export { LegalArchiveService } from './LegalArchiveService';
